+++ app/apptesting/test_suite.go
@@ -311,3 +311,10 @@ func CreateRandomAccounts(numAccts int) []sdk.AccAddress {
 
 	return testAddrs
 }
+
+func GenerateTestAddrs() (string, string) {
+	pk1 := ed25519.GenPrivKey().PubKey()
+	validAddr := sdk.AccAddress(pk1.Address()).String()
+	invalidAddr := sdk.AccAddress("invalid").String()
+	return validAddr, invalidAddr
+}